+++ src/gmail/messages/MessagesMessagePartBody.h
@@ -22,14 +22,14 @@ namespace messages{
     public:
         MessagePartBody(){};
 
-        MessagePartBody(const int64_t& arg){ m_size = arg; };
+        MessagePartBody(const int& arg){ m_size = arg; };
 
     public:
             /**
                 Total number of bytes in the body of the message part.
             */
-        int64_t size()const{return m_size;};
-        MessagePartBody& setSize(const int64_t& arg){m_size=arg;return *this;};
+        int size()const{return m_size;};
+        MessagePartBody& setSize(const int& arg){m_size=arg;return *this;};
 
             /**
                 The body data of a MIME message part. May be empty for MIME
@@ -63,7 +63,7 @@ namespace messages{
             /**
                 Total number of bytes in the body of the message part.
             */
-        int64_t m_size = {0};
+        int m_size = {0};
 
             /**
                 The body data of a MIME message part. May be empty for MIME
